Bugfix UI (#72)
* update topic request

* update

* fix
